matching up interfaces

+++ src/BadGuy.java
@@ -1,5 +1,5 @@
 public interface BadGuy {
-    public void inflictDamage(int dmgInflicted);
+    public void inflictDamage(double dmgInflicted);
     public void reduceDamage(double dmgReduced);
     public void reduceDefense(double defenseReduced);
     public void reduceDodge(double dodgeReduced);
