+++ src/Monster.java
@@ -1,13 +1,13 @@
 public class Monster implements BadGuy {
     String name;
     String type;
-    int hp;
-    int level;
-    int damage;
-    int defense;
-    int dodgeChance;
+    double hp;
+    double level;
+    double damage;
+    double defense;
+    double dodgeChance;
 
-    public Monster(String name, int hp, int level, int damage, int defense, int dodgeChance, String monsterType) {
+    public Monster(String name, double hp, double level, double damage, double defense, double dodgeChance, String monsterType) {
         this.name = name;
         this.hp = hp;
         this.level = level;
@@ -17,19 +17,19 @@ public class Monster implements BadGuy {
         this.type = monsterType;
     }
 
-    public void inflictDamage(int dmg) {
+    public void inflictDamage(double dmg) {
         setHp(getHp() - dmg);
     }
 
-    public void reduceDamage(int dmgReduction) {
+    public void reduceDamage(double dmgReduction) {
         setDamage(getDamage() - dmgReduction);
     }
 
-    public void reduceDefense(int defenseReduction) {
+    public void reduceDefense(double defenseReduction) {
         setDefense(getDefense() - defenseReduction);
     }
 
-    public void reduceDodge(int dodgeReduction) {
+    public void reduceDodge(double dodgeReduction) {
         setDodgeChance(getDodgeChance() - dodgeReduction);
     }
 
@@ -42,46 +42,46 @@ public class Monster implements BadGuy {
         this.name = name;
     }
 
-    public int getHp() {
+    public double getHp() {
         return hp;
     }
 
-    public void setHp(int hp) {
+    public void setHp(double hp) {
         // TODO: a check to make this zero if goes below zero?
         this.hp = hp;
     }
 
-    public int getLevel() {
+    public double getLevel() {
         return level;
     }
 
-    public void setLevel(int level) {
+    public void setLevel(double level) {
         this.level = level;
     }
 
-    public int getDamage() {
+    public double getDamage() {
         return damage;
     }
 
-    public void setDamage(int damage) {
+    public void setDamage(double damage) {
         // TODO error checking, floor to damage of zero?
         this.damage = damage;
     }
 
-    public int getDefense() {
+    public double getDefense() {
         return defense;
     }
 
-    public void setDefense(int defense) {
+    public void setDefense(double defense) {
         // TODO error checking, floor to damage of zero?
         this.defense = defense;
     }
 
-    public int getDodgeChance() {
+    public double getDodgeChance() {
         return dodgeChance;
     }
 
-    public void setDodgeChance(int dodgeChance) {
+    public void setDodgeChance(double dodgeChance) {
         // TODO error checking, floor to damage of zero?
         this.dodgeChance = dodgeChance;
     }
